+++ src/main/java/com/example/auctionserver/adapter/in/web/PostAuctionController.java
@@ -1,0 +1,11 @@
+package com.example.auctionserver.adapter.in.web;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/auctions")
+public class PostAuctionController {
+}
