+++ src/main/java/com/example/auctionserver/application/port/in/GetAuctionUseCase.java
@@ -1,0 +1,8 @@
+package com.example.auctionserver.application.port.in;
+
+import com.example.auctionserver.application.port.out.model.GetAuctionResponse;
+
+public interface GetAuctionUseCase {
+
+    GetAuctionResponse getAuction();
+}
