--- src/main/java/com/example/auctionserver/auction/dto/request/RequestBidDto.java
@@ -1,20 +1,0 @@
-package com.example.auctionserver.auction.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-import static lombok.AccessLevel.PROTECTED;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = PROTECTED)
-@AllArgsConstructor(access = PROTECTED)
-public class RequestBidDto implements Serializable {
-
-    private Long memberId;
-    private Long bid;
-}
