--- src/main/java/com/example/auctionserver/auction/dto/response/ResponseAuctionDto.java
@@ -1,19 +1,0 @@
-package com.example.auctionserver.auction.dto.response;
-
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class ResponseAuctionDto {
-    private Long id;
-    private String productName;
-    private String imageUrl;
-    private Long openingPrice;
-    private LocalDateTime openingTime;
-    private LocalDateTime closingTime;
-    private Long winningPrice;
-}
