--- src/main/java/com/example/auctionserver/auction/dto/response/ResponseWinningPriceDto.java
@@ -1,17 +1,0 @@
-package com.example.auctionserver.auction.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import static lombok.AccessLevel.PROTECTED;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = PROTECTED)
-@AllArgsConstructor(access = PROTECTED)
-public class ResponseWinningPriceDto {
-
-    private Long winningPrice;
-}
