+++ src/main/java/com/example/auctionserver/domain/Product.java
@@ -1,0 +1,33 @@
+package com.example.auctionserver.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+@Entity
+@Table(name = "products")
+@Getter
+@Builder
+@NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor(access = PROTECTED)
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "product_id", nullable = false)
+    private Long id;
+
+    @Column(nullable = false)
+    private Long productId;
+
+    @Column(nullable = false)
+    private String productName;
+
+    @Column(nullable = false)
+    private String imageUrl;
+}
