+++ src/main/java/com/example/auctionserver/global/dto/request/RequestAuctionDto.java
@@ -1,0 +1,18 @@
+package com.example.auctionserver.global.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+import static lombok.AccessLevel.PROTECTED;
+
+@Getter
+@NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor
+public class RequestAuctionDto {
+
+    private Long point;
+    private LocalDateTime time;
+}
