+++ src/main/java/com/example/auctionserver/global/dto/response/ResponsePointDto.java
@@ -1,0 +1,19 @@
+package com.example.auctionserver.global.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import static lombok.AccessLevel.*;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = PROTECTED)
+@AllArgsConstructor(access = PROTECTED)
+public class ResponsePointDto {
+
+    private Long point;
+    private Long deposit;
+    private Long availablePoint;
+}
